requirements.txt
streamlit
python-dotenv
google-generativeai
pymupdf
scikit-learn
numpy
